share progress display code between get and console download
also use pb.set_message so that the progress messages like 1/3 connecting
vanish after the associated step is done
